refactor: reorganize redux selectors folders
