+++ packages/frontend/src/components/staking/components/ClaimTokenFarmRewardsModal.js
@@ -4,7 +4,7 @@ import { useSelector } from 'react-redux';
 import { Textfit } from 'react-textfit';
 import styled from 'styled-components';
 
-import selectNEARAsTokenWithMetadata from '../../../redux/crossStateSelectors/selectNEARAsTokenWithMetadata';
+import selectNEARAsTokenWithMetadata from '../../../redux/selectors/crossStateSelectors/selectNEARAsTokenWithMetadata';
 import FormButton from '../../common/FormButton';
 import Modal from '../../common/modal/Modal';
 import AlertTriangleIcon from '../../svg/AlertTriangleIcon';
