+++ packages/frontend/src/redux/selectors/crossStateSelectors/selectNEARAsTokenWithMetadata.js
